Move the Model account.invoice in a python module with the same name
