+++ account_invoice_reference/__init__.py
@@ -20,3 +20,4 @@
 ##############################################################################
 
 from . import account_move
+from . import account_invoice
